+++ Pipe.py
@@ -1,0 +1,62 @@
+import pygame
+import os
+import random
+
+class Pipe:
+    
+    GAP = 200
+    VEL = 5
+
+    def __init__(self, x):
+        """
+        initialize pipe object
+        :param x: int
+        :param y: int
+        :return" None
+        """
+        self.x = x
+        self.height = 0
+
+        # where the top and bottom of the pipe is
+        self.top = 0
+        self.bottom = 0
+
+        self.PIPE_TOP = pygame.transform.flip(pygame.transform.scale2x(pygame.image.load(os.path.join(os.getcwd(),"res","pipe.png")).convert_alpha()), False, True)
+        self.PIPE_BOTTOM = pygame.transform.scale2x(pygame.image.load(os.path.join(os.getcwd(),"res","pipe.png")).convert_alpha())
+        self.WIDTH = self.PIPE_TOP.get_width()
+
+        self.passed = False
+
+        self.set_height()
+
+    def set_height(self):
+
+        self.height = random.randrange(50, 450)
+        self.top = self.height - self.PIPE_TOP.get_height()
+        self.bottom = self.height + self.GAP
+
+    def move(self):
+
+        self.x -= self.VEL
+
+    def draw(self, win):
+
+        win.blit(self.PIPE_TOP, (self.x, self.top))
+        win.blit(self.PIPE_BOTTOM, (self.x, self.bottom))
+
+
+    def collide(self, bird, win):
+ 
+        bird_mask = bird.get_mask()
+        top_mask = pygame.mask.from_surface(self.PIPE_TOP)
+        bottom_mask = pygame.mask.from_surface(self.PIPE_BOTTOM)
+        top_offset = (self.x - bird.x, self.top - round(bird.y))
+        bottom_offset = (self.x - bird.x, self.bottom - round(bird.y))
+
+        b_point = bird_mask.overlap(bottom_mask, bottom_offset)
+        t_point = bird_mask.overlap(top_mask,top_offset)
+
+        if b_point or t_point:
+            return True
+
+        return False
